+++ src/components/Dropdown/DropdownBlockLinkSecondaryText/index.js
@@ -1,0 +1,31 @@
+import React from 'react';
+import PropTypes from 'prop-types';
+import Icon from 'react-oui-icons';
+
+export default function DropdownBlockLinkSecondaryText(props) {
+  const testSection = props.testSection ? 'block-link-description-' + props.testSection : null;
+  return (
+    <div className="oui-dropdown--descriptive__content flex">
+      { props.isWarning &&
+        <Icon name='exclamation' size='small' />
+      }
+      <div { ...(testSection ? { 'data-test-section': testSection } : {}) }>
+        { props.secondaryText }
+      </div>
+    </div>
+  );
+}
+
+DropdownBlockLinkSecondaryText.defaultProps = {
+  isWarning: false,
+  secondaryText: null,
+};
+
+DropdownBlockLinkSecondaryText.propTypes = {
+  /** should show info icon */
+  isWarning: PropTypes.bool,
+  /** description text, if provided */
+  secondaryText: PropTypes.string,
+  /** test section from parent */
+  testSection: PropTypes.string,
+};